removed redundant code
